Fix compilation failure on DMD 2.67.1
This change fixes the following compilation error:

...\source\toml\parser.d(21): Error: None of the overloads of '__ctor' are callable using argument types (string, string, ulong), candidates are:
c:\D\dmd2\windows\bin\..\..\src\druntime\import\object.di(364):        object.Exception.this(string msg, string file = __FILE__, uint line = cast(uint)__LINE__, Throwable next = null)
 c:\D\dmd2\windows\bin\..\..\src\druntime\import\object.di(369):        object.Exception.this(string msg, Throwable next, string file = __FIL E__, uint line = cast(uint)__LINE__)

+++ source/toml/parser.d
@@ -17,7 +17,7 @@ enum TOMLType {
 };
 
 class TOMLException: Exception {
-    this(string msg, string file="parser", ulong line=111) {
+    this(string msg, string file="parser", uint line=111) {
         super(msg, file, line);
     }
 }
